Tracing functionality added.
Added functions to trace the feeder and populate the __livewire shadow
table. This is the shadow table to end all shadow tables. This table is
essentially a cached set of all possible routes from the source (most
likely the substation transformer) to each end load.

Tracing tests coming up next.

	modified:   livewire--0.2.0.sql
	new file:   sql/lw_redirect.sql
	new file:   sql/lw_sourcenodes.sql
	new file:   sql/lw_traceall.sql
	new file:   sql/lw_tracesource.sql
	modified:   tests/expected/000-functions-exist.out
	modified:   tests/sql/000-functions-exist.sql