Added new variants of lw_tracednstream
Added three new variants of aforemention function. The first allows
tracing from multiple points. The second allows the additions of
blocks to stop the trace. The third allows for the passing of a
point geometry. In this case the nearest node to point will bw used
to initiate the trace down stream.