Changed lw_redirect from a recursive function to a recursive CTE.
This is much more performant (is that evn a word?). It is much
faster and will not run into any max_stack_depth errors.

Also made some changes to NOTICE's for lw_traceall. Changed
lw_tracesource checksource argument to 'False' in lw_traceall
since the single source check is done at the start of that function.